+++ src/modules/vc-helpers/document-loader.js
@@ -5,14 +5,21 @@ export default async function(url) {
   const context = testContext[url];
   if(context) {
     return {
-      contextUrl: null,
       documentUrl: url,
       document: context
     };
   }
+  // TODO: add support for resolving DIDs, maybe something like:
+  // if (url.startsWith('did:')) {
+  //   const did_doc = await resolveEthrDIDAndDockDID(did);
+  //   return {
+  //     contextUrl: null,
+  //     documentUrl: url,
+  //     did_doc
+  //   }
+  // }
   const {data: document} = await axios.get(url);
   return {
-    contextUrl: null,
     document,
     documentUrl: url,
   };
